test: add widget tests for tag filter
