+++ lib/widgets/notes_list.dart
@@ -1,0 +1,162 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:intl/intl.dart';
+import 'package:provider/provider.dart';
+
+import '../models/note.dart';
+import '../providers/note_provider.dart';
+import '../services/auth_service.dart';
+import '../screens/note_detail_screen.dart';
+
+class NotesList extends StatefulWidget {
+  final List<Note> notes;
+
+  const NotesList({super.key, required this.notes});
+
+  @override
+  State<NotesList> createState() => _NotesListState();
+}
+
+class _NotesListState extends State<NotesList> {
+  final ScrollController _scrollController = ScrollController();
+  bool _isLoadingMore = false;
+  bool _hasMore = true;
+  DateTime? _lastFetched;
+  static const _pageSize = 20;
+
+  @override
+  void initState() {
+    super.initState();
+    _scrollController.addListener(_onScroll);
+    WidgetsBinding.instance.addPostFrameCallback((_) => _initialFetch());
+  }
+
+  Future<void> _initialFetch() async {
+    final provider = context.read<NoteProvider>();
+    if (provider.notes.isNotEmpty) {
+      setState(() {
+        _lastFetched = provider.notes.last.updatedAt;
+        _hasMore = provider.notes.length >= _pageSize;
+      });
+    } else {
+      final notes = await provider.fetchNotesPage(null, _pageSize);
+      if (!mounted) return;
+      setState(() {
+        if (notes.isNotEmpty) {
+          _lastFetched = notes.last.updatedAt;
+          _hasMore = notes.length == _pageSize;
+        }
+      });
+    }
+  }
+
+  void _onScroll() {
+    if (!_hasMore || _isLoadingMore) return;
+    if (_scrollController.position.pixels >=
+        _scrollController.position.maxScrollExtent - 200) {
+      _loadMore();
+    }
+  }
+
+  Future<void> _loadMore() async {
+    setState(() => _isLoadingMore = true);
+    final notes = await context.read<NoteProvider>().fetchNotesPage(
+      _lastFetched,
+      _pageSize,
+    );
+    if (!mounted) return;
+    setState(() {
+      _isLoadingMore = false;
+      if (notes.isNotEmpty) {
+        _lastFetched = notes.last.updatedAt;
+      }
+      if (notes.length < _pageSize) {
+        _hasMore = false;
+      }
+    });
+  }
+
+  @override
+  void dispose() {
+    _scrollController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final notes = widget.notes;
+    if (notes.isEmpty) {
+      return Center(child: Text(AppLocalizations.of(context)!.noNotes));
+    }
+
+    final provider = context.watch<NoteProvider>();
+    final itemCount = notes.length + (_isLoadingMore ? 1 : 0);
+    return ListView.builder(
+      controller: _scrollController,
+      itemCount: itemCount,
+      itemBuilder: (context, index) {
+        if (index >= notes.length) {
+          return const Padding(
+            padding: EdgeInsets.symmetric(vertical: 16),
+            child: Center(child: CircularProgressIndicator()),
+          );
+        }
+        final note = notes[index];
+        return Card(
+          child: ListTile(
+            leading: note.locked ? const Icon(Icons.lock) : null,
+            title: Text(note.title),
+            subtitle: Text(
+              note.alarmTime != null
+                  ? '${note.content}\n⏰ ${DateFormat.yMd(Localizations.localeOf(context).toString()).add_Hm().format(note.alarmTime!)}'
+                  : note.content,
+            ),
+            onTap: () async {
+              if (note.locked) {
+                final ok = await AuthService().authenticate(
+                  AppLocalizations.of(context)!,
+                );
+                if (!ok) return;
+              }
+              Navigator.push(
+                context,
+                MaterialPageRoute(builder: (_) => NoteDetailScreen(note: note)),
+              );
+            },
+            trailing: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                if (!provider.isSynced(note.id))
+                  const Icon(Icons.sync_problem, color: Colors.orange),
+                IconButton(
+                  icon: const Icon(Icons.delete),
+                  tooltip: AppLocalizations.of(context)!.delete,
+                  onPressed: () {
+                    final provider = context.read<NoteProvider>();
+                    final idx = provider.notes.indexWhere(
+                      (n) => n.id == note.id,
+                    );
+                    if (idx != -1) {
+                      provider.removeNoteAt(idx);
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            AppLocalizations.of(context)!.noteDeleted,
+                          ),
+                          action: SnackBarAction(
+                            label: AppLocalizations.of(context)!.undo,
+                            onPressed: () => provider.addNote(note),
+                          ),
+                        ),
+                      );
+                    }
+                  },
+                ),
+              ],
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
